requirements.txt
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
idna==2.8
influxdb==5.2.3
iso8601==0.1.12
pyserial==3.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
